restructure the errors
