+++ src/error.rs
@@ -2,10 +2,6 @@
 
 use std::error::Error;
 
-use bitcoin::Amount;
-
-use crate::protocol::error::ContractError;
-
 /// Includes all network-related errors.
 #[derive(Debug)]
 pub enum NetError {
@@ -38,22 +34,5 @@ impl From<std::io::Error> for NetError {
 impl From<serde_cbor::Error> for NetError {
     fn from(value: serde_cbor::Error) -> Self {
         Self::Cbor(value)
-    }
-}
-
-/// Includes all Protocol-level errors.
-#[derive(Debug)]
-pub enum ProtocolError {
-    WrongMessage { expected: String, received: String },
-    WrongNumOfSigs { expected: usize, received: usize },
-    WrongNumOfContractTxs { expected: usize, received: usize },
-    WrongNumOfPrivkeys { expected: usize, received: usize },
-    IncorrectFundingAmount { expected: Amount, found: Amount },
-    Contract(ContractError),
-}
-
-impl From<ContractError> for ProtocolError {
-    fn from(value: ContractError) -> Self {
-        Self::Contract(value)
     }
 }
